Fix BSD License
The content of the license file on the "main" branch looked like someone
had done a "save as" from a GitHub page, so all the surrounding nagivation
HTML junk was included. This replaces all of that with a copy of the text
from the original "master" branch, which is what README.md points directly
to.